Add shift item to it's own module
