+++ packages/mock-server/src/api/resolvers/invoice.ts
@@ -1,0 +1,1 @@
+export const invoice = {};
