+++ packages/mock-server/src/api/resolvers/name.ts
@@ -1,0 +1,1 @@
+export const name = {};
