+++ packages/mock-server/src/api/resolvers/requisition.ts
@@ -1,0 +1,1 @@
+export const requisition = {};
